Praveen|Removing unnecessary cast to int of milliseconds interval

+++ motech-scheduler/src/main/java/org/motechproject/scheduler/MotechSchedulerServiceImpl.java
@@ -278,7 +278,7 @@ public class MotechSchedulerServiceImpl extends MotechObject implements MotechSc
         putMotechEventDataToJobDataMap(jobDetail.getJobDataMap(), motechEvent);
 
         SimpleScheduleBuilder simpleSchedule = simpleSchedule()
-                .withIntervalInMilliseconds((int) repeatIntervalInMilliSeconds)
+                .withIntervalInMilliseconds(repeatIntervalInMilliSeconds)
                 .withRepeatCount(jobRepeatCount);
 
         simpleSchedule = setMisfirePolicyForSimpleTrigger(simpleSchedule, repeatingTriggerMisfirePolicy);
